pyproject.toml
[project]
name = "telegram-assistant-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aisuite[openai]>=0.1.11",
    "docstring-parser>=0.16",
    "google-api-python-client>=2.174.0",
    "google-auth-oauthlib>=1.2.2",
    "openai>=1.93.0",
    "python-telegram-bot>=22.2",
]

[tool.setuptools]
packages = ["tg_bot"]

[project.scripts]
tg-assistant-start = "tg_bot.start:main"
